fixed up cc1 to deal with different firmware versions
